added triangle counting dataset
contains 778 randomly sampled non-isomorphic graphs with 3 to 100 vertices that each contain between 0 and 9 triangles. the goal is to determine the triangle count.

the non-triangle cycle count in the dataset's graphs ranges from 0 (i.e. trees) to approx. 1 cycle per vertex (the upper cycle count bound is not strictly enforced, so it could be lower or higher).