added regression-based sorting evaluation + fixed sorting memory leak
apparently repeated loading and discarding of the same pickle file is leaky in the standard python memory allocator. 

to circumvent this, the "reconfigurable_finalization" flag was added to DatasetProvider and Preprocessor. it allows reuse of the same preprocessed heap data with different batchers (required for the parallel computation of the partition-wise comparisons in the recursive layers of a comparator-model-guided quicksort).

additional benefit: this is also quite a bit faster.